bcdc_single_download_limit returns integer not double

+++ tests/testthat/test-options.R
@@ -43,5 +43,5 @@ test_that("bcdc_single_download_limit returns a number",{
   skip_on_cran()
   skip_if_net_down()
   lt <- bcdc_single_download_limit()
-  expect_type(lt, "double")
+  expect_type(lt, "integer")
 })
